Split package in sdformat_to_mjcf and sdformat_mjcf_utils
