+++ sdformat_mjcf_utils/setup.cfg
