+++ sdformat_mjcf_utils/tox.ini
@@ -1,0 +1,28 @@
+# See https://github.com/pypa/sampleproject/blob/main/tox.ini
+# for the original commented file
+
+[tox]
+envlist = py{37,38,39,310}
+
+minversion = 3.3.0
+isolated_build = true
+
+[testenv]
+# pythonpath has to be passed through to the tox to get python3-ignition-mathX and python3-sdformatX
+passenv = PYTHONPATH
+deps =
+    check-manifest >= 0.42
+    flake8
+    numpy
+    dm_control
+    coverage
+commands =
+    check-manifest --ignore 'tox.ini,tests/**'
+    python setup.py check -m -s
+    python3 -m flake8 .
+    coverage run -m unittest {posargs}
+    coverage xml
+
+[flake8]
+exclude = .tox,*.egg
+select = E,W,F
