+++ sdformat_to_mjcf/sdformat_to_mjcf/__init__.py
@@ -1,4 +1,4 @@
-from sdformat_mjcf.sdformat_mjcf import execute
+from sdformat_to_mjcf.sdformat_mjcf import execute
 
 
 def main():
