+++ sdformat_to_mjcf/sdformat_to_mjcf/__version__.py
@@ -1,0 +1,3 @@
+VERSION = (0, 0, 1)
+
+__version__ = '.'.join(map(str, VERSION))
