+++ sdformat_to_mjcf/sdformat_to_mjcf/converters/joint.py
@@ -18,7 +18,7 @@ import math
 
 import sdformat as sdf
 
-import sdformat_mjcf.sdf_utils as su
+import sdformat_mjcf_utils.sdf_utils as su
 
 JointType = sdf.Joint.JointType
