+++ sdformat_to_mjcf/tests/resources/box_obj/meshes/box.mtl
