+++ sdformat_to_mjcf/tests/test_add_geometry.py
@@ -23,7 +23,7 @@ from ignition.math import Pose3d, Vector2d, Vector3d
 from dm_control import mjcf
 
 from tests import helpers
-from sdformat_mjcf.converters import geometry as geometry_conv
+from sdformat_to_mjcf.converters import geometry as geometry_conv
 
 
 GeometryType = sdf.Geometry.GeometryType
