+++ sdformat_to_mjcf/tests/test_add_link.py
@@ -20,7 +20,7 @@ import sdformat as sdf
 from ignition.math import Inertiald, Pose3d, MassMatrix3d, Vector3d
 from dm_control import mjcf
 
-from sdformat_mjcf.converters.link import add_link
+from sdformat_to_mjcf.converters.link import add_link
 from tests import helpers
 
 GeometryType = sdf.Geometry.GeometryType
