+++ sdformat_to_mjcf/tox.ini
@@ -4,7 +4,7 @@
 [tox]
 envlist = py{37,38,39,310}
 
-minversion = 3.7.0
+minversion = 3.3.0
 isolated_build = true
 
 [testenv]
@@ -13,9 +13,8 @@ passenv = PYTHONPATH
 deps =
     check-manifest >= 0.42
     flake8
-    numpy
-    dm_control
     coverage
+    ../sdformat_mjcf_utils
 commands =
     check-manifest --ignore 'tox.ini,tests/**'
     python setup.py check -m -s
